Changes leaflet import to * for esm version
On Leaflet’s master branch they’ve added a "module" entry that points to
a modulized version of Leaflet. Since that module has no default export,
this package needs to change to import from 'leaflet'.

Changed to import just pieces to be supportive of tree-shaking rather
than pulling in all of Leaflet.